Apply Facade Pattern to MusicPlayCommand and PlayerManagerService(Facade) for reducing complexity in MusicPlayCommand dependencies
