+++ src/main/java/com/discord/bot/commands/musiccommands/QueueCommand.java
@@ -2,14 +2,8 @@ package com.discord.bot.commands.musiccommands;
 
 import com.discord.bot.commands.musiccommands.Fails.FailQueueEmptyStrategy;
 import com.discord.bot.service.audioplayer.PlayerManagerService;
-import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
-import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.BlockingQueue;
 
 public class QueueCommand extends MusicPlayerCommand {
 
@@ -18,31 +12,16 @@ public class QueueCommand extends MusicPlayerCommand {
         this.channelValidation = channelValidation;
         this.failDescriptionStrategy = new FailQueueEmptyStrategy();
     }
-
-
+    
     @Override
     void operate(SlashCommandInteractionEvent event, EmbedBuilder embedBuilder) {
-        BlockingQueue<AudioTrack> queue = playerManagerService.getMusicManager(event).scheduler.queue;
-        int trackCount = Math.min(queue.size(), 20);
-        List<AudioTrack> trackList = new ArrayList<>(queue);
-
-        embedBuilder.setTitle("Current Queue:");
-        for (int i = 0; i < trackCount; i++) {
-            AudioTrack track = trackList.get(i);
-            AudioTrackInfo info = track.getInfo();
-            embedBuilder.appendDescription((i + 1) + ". " + info.title + "\n");
-        }
-
-        if (trackList.size() > trackCount) {
-            embedBuilder.appendDescription("And " + (trackList.size() - trackCount) + " more...");
-        }
-
-        event.replyEmbeds(embedBuilder.build()).queue();
+        playerManagerService.showTrackQueue(event, embedBuilder);
     }
-
+    
     @Override
     boolean isValidState(SlashCommandInteractionEvent event) {
-        BlockingQueue<AudioTrack> queue = playerManagerService.getMusicManager(event).scheduler.queue;
-        return !queue.isEmpty();
+        return playerManagerService.isQueueNotEmpty(event);
     }
+
+    
 }
